Application controller was moved into the library module. Now It can be reused in upper level application.
